App, WiseSayingController 책임분리
