+++ src/main/java/org/example/domain/wiseSaying/App.java
@@ -1,6 +1,7 @@
 package org.example.domain.wiseSaying;
 
 import org.example.domain.wiseSaying.controller.SystemController;
+import org.example.domain.wiseSaying.controller.WiseSayingController;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -11,8 +12,8 @@ public class App {
     private Scanner sc = new Scanner(System.in);
     private int lastId = 0;
     private List<WiseSaying> wiseSayings = new ArrayList<>();
-    private SystemController  systemController = new SystemController();
-
+    private SystemController systemController = new SystemController();
+    private WiseSayingController wiseSayingController = new WiseSayingController(sc);
 
     public void run() {
 
@@ -25,18 +26,16 @@ public class App {
             Rq rq = new Rq(command);
             String actionName = rq.getActionName();
             if (actionName.equals("등록")) {
-                actionWrite();
+                wiseSayingController.actionWrite();
 
             } else if (actionName.equals("목록")) {
-                actionList();
+                wiseSayingController.actionList();
 
             } else if (actionName.startsWith("삭제")) {
-                actionDelete(rq);
-
+                wiseSayingController.actionDelete(rq);
 
             } else if (actionName.startsWith("수정")) {
-                actionModify(rq);
-
+                wiseSayingController.actionModify(rq);
 
             } else if (actionName.equals("종료")) {
                 systemController.exit();
@@ -48,88 +47,4 @@ public class App {
     }
 
 
-
-    private void actionModify(Rq rq) {
-
-        int id = rq.getParamAsInt("id", -1);
-        WiseSaying wiseSaying = findByIdOrNull(id);
-
-        if(wiseSaying == null) {
-            System.out.println("%d번 명언은 존재하지 않습니다.".formatted(id));
-            return;
-        }
-
-        System.out.println("명언(기존) : %s".formatted(wiseSaying.getSaying()));
-        System.out.print("명언 : ");
-        String newSaying = sc.nextLine();
-        System.out.println("작가(기존) : %s".formatted(wiseSaying.getAuthor()));
-        System.out.print("작가 : ");
-        String newAuthor = sc.nextLine();
-
-        modify(wiseSaying, newSaying, newAuthor);
-    }
-
-    private void modify(WiseSaying wiseSaying, String newSaying, String newAuthor) {
-        wiseSaying.setSaying(newSaying);
-        wiseSaying.setAuthor(newAuthor);
-    }
-
-    private void actionDelete(Rq rq) {
-
-        int id = rq.getParamAsInt("id", -1);
-
-        boolean result = delete(id);
-
-        if (result) {
-            System.out.println("%d번 명언이 삭제되었습니다.".formatted(id));
-        } else {
-            System.out.println("%d번 명언은 존재하지 않습니다.".formatted(id));
-        }
-    }
-
-    private WiseSaying findByIdOrNull(int id) {
-        return wiseSayings.stream()
-                .filter(w -> w.getId() == id)
-                .findFirst()
-                .orElse(null);
-    }
-
-    private boolean delete(int id) {
-        return wiseSayings.removeIf(w -> w.getId() == id);
-    }
-
-    private void actionList() {
-        System.out.println("번호 / 작가 / 명언");
-        System.out.println("----------------------");
-
-        List<WiseSaying> wiseSayings = findListDesc();
-
-        for (WiseSaying wiseSaying : wiseSayings) {
-            System.out.println("%d / %s / %s".formatted(wiseSaying.getId(), wiseSaying.getSaying(), wiseSaying.getAuthor()));
-        }
-    }
-
-    private List<WiseSaying> findListDesc() {
-        return wiseSayings.reversed();
-    }
-
-    private void actionWrite() {
-        System.out.print("명언 : ");
-        String saying = sc.nextLine();
-        System.out.print("작가 : ");
-        String author = sc.nextLine();
-
-        WiseSaying wiseSaying = write(saying, author);
-
-        System.out.println("%d번 명언이 등록되었습니다.".formatted(wiseSaying.getId()));
-    }
-
-    private WiseSaying write(String saying, String author) {
-
-        lastId++;
-        WiseSaying wiseSaying = new WiseSaying(lastId, saying, author);
-        wiseSayings.add(wiseSaying);
-
-        return wiseSaying;
-    }
 }
