+++ src/main/java/org/example/domain/wiseSaying/controller/WiseSayingController.java
@@ -1,4 +1,103 @@
 package org.example.domain.wiseSaying.controller;
 
+import org.example.domain.wiseSaying.Rq;
+import org.example.domain.wiseSaying.WiseSaying;
+
+import java.util.ArrayList;
+import java.util.List;
+import java.util.Scanner;
+
 public class WiseSayingController {
+    private int lastId = 0;
+    private List<WiseSaying> wiseSayings = new ArrayList<>();
+    private Scanner sc;
+
+    public WiseSayingController(Scanner sc) {
+        this.sc = sc;
+    }
+
+    public void actionModify(Rq rq) {
+
+        int id = rq.getParamAsInt("id", -1);
+        WiseSaying wiseSaying = findByIdOrNull(id);
+
+        if (wiseSaying == null) {
+            System.out.println("%d번 명언은 존재하지 않습니다.".formatted(id));
+            return;
+        }
+
+        System.out.println("명언(기존) : %s".formatted(wiseSaying.getSaying()));
+        System.out.print("명언 : ");
+        String newSaying = sc.nextLine();
+        System.out.println("작가(기존) : %s".formatted(wiseSaying.getAuthor()));
+        System.out.print("작가 : ");
+        String newAuthor = sc.nextLine();
+
+        modify(wiseSaying, newSaying, newAuthor);
+    }
+
+    private void modify(WiseSaying wiseSaying, String newSaying, String newAuthor) {
+        wiseSaying.setSaying(newSaying);
+        wiseSaying.setAuthor(newAuthor);
+    }
+
+    public void actionDelete(Rq rq) {
+
+        int id = rq.getParamAsInt("id", -1);
+
+        boolean result = delete(id);
+
+        if (result) {
+            System.out.println("%d번 명언이 삭제되었습니다.".formatted(id));
+        } else {
+            System.out.println("%d번 명언은 존재하지 않습니다.".formatted(id));
+        }
+    }
+
+    private WiseSaying findByIdOrNull(int id) {
+        return wiseSayings.stream()
+                .filter(w -> w.getId() == id)
+                .findFirst()
+                .orElse(null);
+    }
+
+    private boolean delete(int id) {
+        return wiseSayings.removeIf(w -> w.getId() == id);
+    }
+
+    public void actionList() {
+        System.out.println("번호 / 작가 / 명언");
+        System.out.println("----------------------");
+
+        List<WiseSaying> wiseSayings = findListDesc();
+
+        for (WiseSaying wiseSaying : wiseSayings) {
+            System.out.println("%d / %s / %s".formatted(wiseSaying.getId(), wiseSaying.getSaying(), wiseSaying.getAuthor()));
+        }
+    }
+
+    private List<WiseSaying> findListDesc() {
+        return wiseSayings.reversed();
+    }
+
+    public void actionWrite() {
+        System.out.print("명언 : ");
+        String saying = sc.nextLine();
+        System.out.print("작가 : ");
+        String author = sc.nextLine();
+
+        WiseSaying wiseSaying = write(saying, author);
+
+        System.out.println("%d번 명언이 등록되었습니다.".formatted(wiseSaying.getId()));
+    }
+
+    private WiseSaying write(String saying, String author) {
+
+        lastId++;
+        WiseSaying wiseSaying = new WiseSaying(lastId, saying, author);
+        wiseSayings.add(wiseSaying);
+
+        return wiseSaying;
+    }
+
 }
